requirements.txt
python-telegram-bot==20.7
pandas==2.1.4
openpyxl==3.1.2
python-dotenv==1.0.0
pytz==2023.3
ulid-py==1.1.0
